Changes to be committed: Ultimos metodos agregados en los moodelos Paciente y Doctor,     y se agrego validaciones extra para la verificación de los roles
	modified:   API-Clinica/Controllers/AdminDoctorController.cs
	modified:   API-Clinica/Controllers/AdminPatientController.cs
	modified:   API-Clinica/Controllers/AdministratorController.cs
	modified:   API-Clinica/Controllers/AppointmentController.cs
	modified:   API-Clinica/Controllers/DoctorController.cs
	modified:   API-Clinica/Controllers/PatientController.cs
	modified:   API-Clinica/Models/Specialty.cs
	modified:   API-Clinica/Services/DoctorDbService.cs
	modified:   API-Clinica/bin/Debug/net8.0/API-Clinica.dll
	modified:   API-Clinica/bin/Debug/net8.0/API-Clinica.exe
	modified:   API-Clinica/bin/Debug/net8.0/API-Clinica.pdb
	modified:   API-Clinica/obj/Debug/net8.0/API-Clinica.AssemblyInfo.cs
	modified:   API-Clinica/obj/Debug/net8.0/API-Clinica.AssemblyInfoInputs.cache
	modified:   API-Clinica/obj/Debug/net8.0/API-Clinica.dll
	modified:   API-Clinica/obj/Debug/net8.0/API-Clinica.pdb
	modified:   API-Clinica/obj/Debug/net8.0/API-Clinica.sourcelink.json
	modified:   API-Clinica/obj/Debug/net8.0/apphost.exe
	modified:   API-Clinica/obj/Debug/net8.0/ref/API-Clinica.dll
	modified:   API-Clinica/obj/Debug/net8.0/refint/API-Clinica.dll

+++ API-Clinica/Controllers/AdminDoctorController.cs
@@ -3,68 +3,83 @@ using Microsoft.AspNetCore.Mvc;
 
 [ApiController]
 [Route("api/adm_doctor")]
-public class AdminDoctorController : ControllerBase{
+public class AdminDoctorController : ControllerBase
+{
 
     private readonly IDoctorService _doctorService;
     private readonly AccountDbService _accountService;
-    public AdminDoctorController(IDoctorService doctorService, AccountDbService accountService){
+    public AdminDoctorController(IDoctorService doctorService, AccountDbService accountService)
+    {
         this._doctorService = doctorService;
         this._accountService = accountService;
     }
 
 
-    [Authorize(Roles = "administrator,Administrator,ADMINISTRATOR")]
+    //[Authorize(Roles = "administrator,Administrator,ADMINISTRATOR")]
 
 
     //[ApiExplorerSettings(IgnoreApi = true)]
     [HttpGet("doctors")]
+    [Authorize(Roles = "administrator,Administrator,ADMINISTRATOR")]
     public ActionResult<List<Doctor>> GetAll(){ // Obtiene todos los Doctores de la BD
         //Falta mostrar las especialidades de los medicos
-        try
-      {
-        return Ok(_doctorService.GetAll());
-      }
-      catch (System.Exception e)
-      {
-        Console.WriteLine(e.Message);
-        return Problem(detail: e.Message, statusCode: 500);
-      }
+        if (!User.IsInRole("administrator") && !User.IsInRole("Administrator") && !User.IsInRole("ADMINISTRATOR")) return Forbid();
+        try{
+            return Ok(_doctorService.GetAll());
+        }
+        catch (System.Exception e){
+            Console.WriteLine(e.Message);
+            return Problem(detail: e.Message, statusCode: 500);
+        }
     }
 
 
     //[ApiExplorerSettings(IgnoreApi = true)]
     //[Authorize(Roles = "admin")]
     [HttpGet("doctor/{id}")]
-    public ActionResult<Doctor> GetById(int id){ //Obtengo un Doctor por su ID
+    [Authorize(Roles = "administrator,Administrator,ADMINISTRATOR")]
+    public ActionResult<Doctor> GetById(int id)
+    { //Obtengo un Doctor por su ID
+        if (!User.IsInRole("administrator") && !User.IsInRole("Administrator") && !User.IsInRole("ADMINISTRATOR")) return Forbid();
         Doctor? _doctor = _doctorService.GetById(id);
 
-        if(_doctor == null) return NotFound("Doctor not found");
+        if (_doctor == null) return NotFound("Doctor not found");
         else return Ok(_doctor);
     }
-    
+
 
     //[ApiExplorerSettings(IgnoreApi = true)]
     //[Authorize(Roles = "admin")]
+
     [HttpPost("doctor")]
-    public ActionResult<Doctor> Create([FromBody] DoctorDTO doctor){
-        if(doctor == null) return BadRequest("Doctor data is required.");
-        if(!ModelState.IsValid) return BadRequest(ModelState); // retorna 400 si hay errores en el modelo
+    [Authorize(Roles = "administrator,Administrator,ADMINISTRATOR")]
+    public ActionResult<Doctor> Create([FromBody] DoctorDTO doctor)
+    {
+        // Verificar que el usuario tenga el rol adecuado
+        if (!User.IsInRole("administrator") && !User.IsInRole("Administrator") && !User.IsInRole("ADMINISTRATOR")) return Forbid(); // Devuelve un código 403 Forbidden si el usuario no tiene permiso
+
+        if (doctor == null) return BadRequest("Doctor data is required.");
+        if (!ModelState.IsValid) return BadRequest(ModelState); // retorna 400 si hay errores en el modelo
 
         Doctor _doctor = _doctorService.Create(doctor); // Creo al nuevo doctor
 
-        if(_doctor == null) return BadRequest("Error creating doctor.");
+        if (_doctor == null) return BadRequest("Error creating doctor.");
 
-        return CreatedAtAction(nameof(GetById), new {id = _doctor.Id}, _doctor); // Devuelvo el resultado de llamar al metodo GetById pasando como parametro el Id del nuevo doctor
+        return CreatedAtAction(nameof(GetById), new { id = _doctor.Id }, _doctor); // Devuelvo el resultado de llamar al metodo GetById pasando como parametro el Id del nuevo doctor
     }
 
 
     //[ApiExplorerSettings(IgnoreApi = true)]
     //[Authorize(Roles = "admin")]
     [HttpDelete("doctor/{id}")]
-    public ActionResult DeleteById(int id){
+    [Authorize(Roles = "administrator,Administrator,ADMINISTRATOR")]
+    public ActionResult DeleteById(int id)
+    {
+        if (!User.IsInRole("administrator") && !User.IsInRole("Administrator") && !User.IsInRole("ADMINISTRATOR")) return Forbid();
+
         var _doctor = _doctorService.GetById(id);
 
-        if(_doctor == null) return NotFound("Doctor not found!!!");
+        if (_doctor == null) return NotFound("Doctor not found!!!");
 
         _doctorService.Delete(id);
         return NoContent();
@@ -72,15 +87,17 @@ public class AdminDoctorController : ControllerBase{
 
 
     [HttpPut("doctor/{id}")]
-    //[Authorize(Roles = "admin")]
-    public ActionResult<Doctor> UpdateById(int id, DoctorDTO d){
+    [Authorize(Roles = "administrator,Administrator,ADMINISTRATOR")]
+    public ActionResult<Doctor> UpdateById(int id, DoctorDTO d)
+    {
+        if (!User.IsInRole("administrator") && !User.IsInRole("Administrator") && !User.IsInRole("ADMINISTRATOR")) return Forbid();
         try{
 
             Doctor? doctor = _doctorService.Update(id, d);
 
-            if(doctor is null) return NotFound(new {Message = $"No se pudo actualizar el doctor con id: {id}"});
+            if (doctor is null) return NotFound(new { Message = $"No se pudo actualizar el doctor con id: {id}" });
 
-            return CreatedAtAction(nameof(GetById), new{id = doctor.Id}, doctor);
+            return CreatedAtAction(nameof(GetById), new { id = doctor.Id }, doctor);
         }
         catch (Exception e){
             Console.WriteLine(e.Message);
